requirements.txt
langchain==0.2.16
langchain-core==0.2.38
langchain-community==0.2.16

langchain-groq
langchain-pinecone

flask==3.1.1
python-dotenv==1.1.0

sentence-transformers==2.7.0
pypdf==4.2.0

pinecone-client==3.2.2